requirements.txt
# Audio processing
librosa>=0.10.0
numpy>=1.20.0
soundfile>=0.10.0
scipy>=1.7.0
pyacoustid>=1.2.0  # For audio fingerprinting and song identification

# Web server
flask>=2.0.0
requests>=2.25.0

# Recommended but optional
# ffmpeg  # System dependency for audio conversion
